# organs/rosie/packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: rosie-api
  version: 1.0.0
  summary: The execution and observability API behind the rosie widget and the rosie admin console.
  description: |
    The execution and observability API behind the rosie widget and the rosie
    admin console. rosie-api powers both the floating rosie widget (embedded in
    a11oy / amaru / sentra / vessels) and the standalone rosie-operator-console
    Gradio Space.

    Framing: rosie commands a11oy, which enforces; a11oy delegates to amaru and
    sentra; vessels carries the load.

    It exposes three capability classes:
      * read queries (receipts, mesh health, doctrine sweep, live formulas)
      * a question/answer surface (`/v1/ask`) that records an audit receipt
      * a two-phase, human-confirmed execution surface (`/v1/execute` then
        `/v1/execute/confirm`)

    All write paths emit DSSE receipts via a11oy's receipt substrate. The
    canonical DSSE v1 PAE is used for all signatures:
    `"DSSEv1" SP LEN(type) SP type SP LEN(body) SP body`
    (see PhD Crypto verdict finding A1).
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: SZL Holdings Engineering
    url: https://docs.szlholdings.com
servers:
  - url: https://rosie-api.szlholdings.com
    description: UDS Core production gateway
...
security:
  - oidcBearer: []
paths:
...
  /v1/receipts/{id}:
    get:
      tags: [receipts]
      summary: Fetch a single receipt with full DSSE envelope and chain pointers.
      operationId: getReceipt
      security:
        - oidcBearer: [szl-operators]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The receipt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    oidcBearer:
      type: openIdConnect
      description: |
        OIDC Bearer (Keycloak via UDS authservice), same realm as
        szl-receipts-server. The JWT must carry:
          * `sub`   — subject (operator or agent identity)
          * `aud`   — must equal `rosie-api`
          * `groups` — must include `szl-operators` for read endpoints and
                       `szl-executors` for `/v1/execute*`
        `/v1/execute/confirm` additionally requires a `step_up_completed_at`
        claim (ISO-8601 or epoch seconds) less than 5 minutes old.
      openIdConnectUrl: https://sso.szlholdings.com/realms/szl/.well-known/openid-configuration
  responses:
    Unauthorized:
      description: Missing or invalid bearer token (problem+json).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    NotFound:
      description: Resource not found (problem+json).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    DsseSignature:
      type: object
      required: [keyid, sig, alg]
      properties:
        keyid:
          type: string
        sig:
          type: string
          description: base64url-encoded signature over the canonical DSSE v1 PAE.
        alg:
          type: string
          enum: [ed25519]
    ChainPointer:
      type: object
      required: [prev_hash, index, ts]
      properties:
        prev_hash:
          type: string
          description: SHA-256 of the previous receipt line, or "GENESIS".
        index:
          type: integer
          minimum: 0
        ts:
          type: string
          format: date-time
    Receipt:
      type: object
      description: |
        Full DSSE envelope plus SZL chain metadata. Cross-references
        @szl-holdings/a11oy-receipt-substrate types.
      required: [payload, payloadType, signatures, chain]
      properties:
        payload:
          type: string
          description: base64 of the canonical-JSON receipt body.
        payloadType:
          type: string
          examples: ["application/vnd.szl.receipt.v1+json"]
        signatures:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/DsseSignature"
        chain:
          $ref: "#/components/schemas/ChainPointer"
...
    Problem:
      type: object
      description: RFC 7807 problem details.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: A URI under https://docs.szlholdings.com/errors/.
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string